Bookmark button moved to separate component
