+++ project/src/components/offer-card/offer-card.tsx
@@ -1,12 +1,10 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Link } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
-import { adaptOfferToFront } from '../../utils/adapters';
-import { api } from '../../index';
-import { APIRoute, AppRoute, IsFavoriteValue } from '../../const';
-import browserHistory from '../../services/browser-history';
+import { AppRoute } from '../../const';
 import { getVisualRating } from '../../utils/common-utils';
 import { getOfferDataAction } from '../../store/api-actions';
+import OfferCardBookmark from '../offer-card-bookmark/offer-card-bookmark';
 import { OfferDataTypes } from '../../types/offer-data-types';
 
 type OfferCardTypes = {
@@ -17,27 +15,11 @@ type OfferCardTypes = {
 };
 
 function OfferCardConnected(props: OfferCardTypes): JSX.Element {
+  const dispatch = useDispatch();
   const { data: offerData, onActiveCardChange, articleClass, imgWrapperClass } = props;
   const { id, price, rating, title, previewImage, type, isPremium, isFavorite } = offerData;
 
-  const dispatch = useDispatch();
-
   const visualRating = getVisualRating(rating);
-
-  const [isFavoriteStatus, setIsFavoriteStatus] = useState(isFavorite);
-
-  const bookmarkButtonClass = isFavoriteStatus
-    ? 'place-card__bookmark-button place-card__bookmark-button--active button'
-    : 'place-card__bookmark-button button';
-
-  const handleBookmarkClick = async (hotelId: string): Promise<void> => {
-    const isFavoriteValue = isFavoriteStatus ? IsFavoriteValue.NotFavorite : IsFavoriteValue.Favorite;
-    await api.post(`${ APIRoute.Favorite }/${ hotelId }/${ isFavoriteValue }`)
-      .then(({ data }) => {
-        setIsFavoriteStatus(adaptOfferToFront(data).isFavorite);
-      })
-      .catch(() => browserHistory.push(AppRoute.Login));
-  };
 
   return (
     <article className={ `${ articleClass } place-card` }
@@ -66,16 +48,10 @@ function OfferCardConnected(props: OfferCardTypes): JSX.Element {
             <b className="place-card__price-value">&euro;{ price }</b>
             <span className="place-card__price-text">&#47;&nbsp;night</span>
           </div>
-          <button
-            className={ bookmarkButtonClass }
-            type="button"
-            onClick={ () => handleBookmarkClick(id) }
-          >
-            <svg className="place-card__bookmark-icon" width="18" height="19">
-              <use xlinkHref="#icon-bookmark" />
-            </svg>
-            <span className="visually-hidden">To bookmarks</span>
-          </button>
+          <OfferCardBookmark
+            isFavorite={ isFavorite }
+            offerId={ id }
+          />
         </div>
         <div className="place-card__rating rating">
           <div className="place-card__stars rating__stars">
